requirements/production.txt
psycopg2==2.6.1
python-dateutil==2.4.2
Babel==1.3
Django==1.8.3
django-floppyforms==1.4.1
